modularize utilities
