+++ templates/luma/dayjs.ts
@@ -1,0 +1,6 @@
+import Dayjs from 'dayjs'
+import LocalizedFormat from 'dayjs/plugin/localizedFormat'
+
+Dayjs.extend(LocalizedFormat)
+
+export const dayjs = Dayjs
